% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/library_list.R
\name{library_list}
\alias{library_list}
\title{Regular Expression Minimal .rda Library}
\usage{
library_list(path, out = NULL, ...)
}
\arguments{
\item{path}{The path to the regular expression library package.}

\item{out}{The directory to output the  documents.}

\item{\ldots}{Other arguments passed to internal functions.}
}
\value{
Generates a minimal .rda regular expression library.
}
\description{
Generate a minimal .rda regular expression library.
}
\examples{
\dontrun{
library_list(system.file("sample", package = "regextools"),
    document = FALSE, install = FALSE)
}
}
\seealso{
\code{\link[base]{save}}
}

