% Generated by roxygen2 (4.0.2): do not edit by hand
\name{library_template}
\alias{library_template}
\title{Regular Expression Library Package Template}
\usage{
library_template(path, name = getOption("name"), email = getOption("email"),
  news = TRUE, readme = TRUE, rstudio = TRUE, gitignore = TRUE,
  testthat = TRUE, travis = TRUE, coverage = TRUE,
  github.user = getOption("github.user"))
}
\arguments{
\item{path}{location to create new regular expression library package. The
last component of the path will be used as the package name.}

\item{name}{A named vector that minimally contains the user's first and last
name (e.g., \code{c(first="Tyler", last="Rinker"))}).  This can be set in the
user's \code{options} in the \file{.Rprofile}; for example:
\code{options(name = c(first="Tyler", middle = "W.", last="Rinker"))}.}

\item{email}{An email address to use for \bold{CRAN} maintainer.  This can be
set in the user's \code{options} in the \file{.Rprofile}; for example:
\code{options(email = "tyler.rinker@gmail.com")}.}

\item{news}{logical.  If \code{TRUE} a \file{NEWS} file is generated.}

\item{readme}{logical.  If code{TRUE} a \file{README.md} file is generated.}

\item{rstudio}{logical.  If code{TRUE} it is assumed
\href{http://www.rstudio.com/}{RStudio} will be used and a \file{xxx.proj}
file is generated.}

\item{gitignore}{logical.  If code{TRUE} a \file{.gitignore} file is generated.}

\item{testthat}{logical.  If code{TRUE} it is assumed
\href{http://journal.r-project.org/archive/2011-1/RJournal_2011-1_Wickham.pdf}{\pkg{testthat}}
will be used and a \file{test} subfolder with appropriate \pkg{testthat}
subdirectories and files will be created.}

\item{travis}{logical.  If code{TRUE} it is assumed
\href{https://travis-ci.org}{Travis-CI} will be used and a \file{travis.yml}
file is generated.  For more on managing a \file{travis.yml} with
\pkg{R} see: \url{https://github.com/craigcitro/r-travis}.}

\item{coverage}{logical.  If code{TRUE} it is assumed
\href{https://github.com/jimhester/covr}{\pkg{covr}} will be used.  This
information will be added to the \file{travis.yml}.}

\item{github.user}{The user's \href{https://github.com/}{GitHub} user name.
This can be set in the user's \code{options} in the \file{.Rprofile}; for example:
\code{options(github.user = "trinker")}.}
}
\description{
This function creates a package template to manage regular expressions.  It
utiizes a framework with defaults that expects the user to use \pkg{testthat}
to test the library's reguar expressions.
}
\examples{
\donttest{
library_template("DELETE_ME")
}
}
\keyword{template}

