% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/library_vignette.R
\name{library_vignette}
\alias{library_vignette}
\title{Regular Expression Library Documentation}
\usage{
library_vignette(path, out = file.path(path, "vignette"),
  is.vignette = TRUE, include.html = !is.vignette, include.md = FALSE,
  theme = "cerulean", iframe = FALSE, ...)
}
\arguments{
\item{path}{The path to the regular expression library package.}

\item{out}{The directory to output the  documents.}

\item{is.vignette}{logical.  Is this an actual package vignette or rendered
html/md reference document.}

\item{include.html}{logical.  If \code{TRUE} an .html rendering of the .Rmd
file is generated.}

\item{include.md}{logical.  If \code{TRUE} a .md rendering of the .Rmd file
is generated.}

\item{theme}{A valid Bootstrap theme (see:
\href{http://rmarkdown.rstudio.com/html_document_format.html#appearance-and-style}{RStudio Themes}).}

\item{iframe}{logical If \code{TRUE} the \href{https://www.debuggex.com/}{Debuggex}
is included as an iframe rather than a link.}

\item{\ldots}{Other arguments passed to internal functions.}
}
\value{
Generates a .Rmd document and optionally .html/.md versions.
}
\description{
Generate a regular expression library package vignette or Rmd/html/md document.
}
\examples{
\dontrun{
library_vignette(system.file("sample", package = "regextools"), "vignette",
    is.vignette = FALSE, include.html=TRUE, document = FALSE, install = FALSE)
}
}
\seealso{
\code{\link[rmarkdown]{render}}
}

